requirements.txt
dash
plotly
PyQt5
numpy
opencv-python
pyyaml
open3d
matplotlib
natsort
imageio
utm